Batman installation
Batman guide will help to install batman and get the node to work as a mesh network. batman_setup.sh has all the command lines to install batman, set ip, and also assign the node to the mesh network so that can be detected by other nodes in the network. batman.service is a daemon file so that everytime the system reboots we don't have to run all the commands in batman-setup.sh file. The daemon will automatically run it and you can use "sudo batctl n" so see what nodes are being detected.